Use install lists to pass dependencies to docker builds
One behavioral change to note here is that the resolved package
versions aren't currently used at install time, and are just printed
for debugging purposes.

With this change, apt will attempt to specifically install the
version which was resolved at generation time.

The primary motivation for this change is to support the use case
where the dependencies aren't known at dockerfile generation time.