Use Colcon extensions to do some neat stuff
Since Colcon is now enumerating the packages, we don't have to skip
all of the non-ros keys when resolving with rosdep. Somewhere in this
change, a significant portion of the test dependencies were corrected,
cutting down on the 'workaround' package installs. Still need to track
those down.

Also, the `build_ignore` values are now package names instead.