More work toward proper rosdep detection
Detecting the rosdeps in a Colcon source space is straightforward
because we know the package type (non-ROS packages' external
dependencies might not be rosdep keys), and we know the dependencies.

For installed packages, I know of two strategies:
1. Look for packages under `install_isolated/share`. This works OK, but
   doesn't detect a fair amount of the installed packages and breaks
   when detecting `ament_python` packages, requiring a nasty hack to
   extract the dependencies for those packages. This has the advantage
   of working right on a Catkin install space as well.
2. Using the colcon index (if present), we can get the full list of
   packages and their deps, but not their types, which means we can't
   use the deps at all.

The approach taken here combines both of these methods. It assumes that
(1) will catch all of the ROS packages and therefore their rosdeps, and
that (2) will enumerate any non-ROS packages which are installed, so
that they can be filtered out of the list of dependencies. This should
leave only rosdeps left in the dependency list.